+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -278,7 +278,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KZ1FPKRAH25MW7D20CVMX8S0"),
+  _id: z.string().default("01KZ8NHRBYWCXRTY5ATH1V23NX"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,