+++ packages/orq-rc/src/models/operations/modelcreatevertex.ts
@@ -44,7 +44,7 @@ export type ModelCreateVertexResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
-  sharing?: components.Config | undefined;
+  sharing?: components.ModelSharingConfig | undefined;
   updated: string;
 };
 
@@ -112,7 +112,7 @@ export const ModelCreateVertexResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
-  sharing: components.Config$inboundSchema.optional(),
+  sharing: components.ModelSharingConfig$inboundSchema.optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {