+++ packages/orq-rc/src/models/operations/modelupdate.ts
@@ -57,7 +57,7 @@ export type ModelUpdateResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
-  sharing?: components.Config | undefined;
+  sharing?: components.ModelSharingConfig | undefined;
   updated: string;
 };
 
@@ -166,7 +166,7 @@ export const ModelUpdateResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
-  sharing: components.Config$inboundSchema.optional(),
+  sharing: components.ModelSharingConfig$inboundSchema.optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {