+++ packages/orq-rc/src/models/operations/modelupdateawsbedrock.ts
@@ -72,7 +72,7 @@ export type ModelUpdateAwsBedrockResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
-  sharing?: components.Config | undefined;
+  sharing?: components.ModelSharingConfig | undefined;
   updated: string;
 };
 
@@ -228,7 +228,7 @@ export const ModelUpdateAwsBedrockResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
-  sharing: components.Config$inboundSchema.optional(),
+  sharing: components.ModelSharingConfig$inboundSchema.optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {