+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -187,7 +187,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KZ1FPKRCWNB8WDWJKZ414GDV"),
+  _id: z.string().default("01KZ8NHRC05P68XD8S32H28NS6"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,