+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-02T14:57:45.633Z",
+    "2026-08-05T09:54:35.619Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {