+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -221,7 +221,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KZ1FPKREY8KY2FBJGKWD7463"),
+  _id: z.string().default("01KZ8NHRC3HGENPHD85X6M2D9N"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,